pyproject.toml
[project]
name = "hello-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.4.2",
]

[dependency-groups]
dev = [
    "dotenv>=0.9.9",
]
